+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change log
 
+## [0.1.5] - 2023-01-26
+### Added
+- Class documentation on manual mode
+- Manual mode section added in README.md
+
+### Modififed
+- Manual mode modified
+
+### Fixed
+
 ## [0.1.4] - 2023-01-25
 ### Added
 - PID class documentation